# ui/openapi.json
{
  "components": {
    "schemas": {
...
      "ClaimIn": {
        "properties": {
          "analytic_confidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analytic Confidence"
          },
          "assertion_type": {
            "default": "reported",
            "title": "Assertion Type",
            "type": "string"
          },
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "collection_method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collection Method"
          },
          "credibility_normalized": {
            "default": "cannot_judge",
            "title": "Credibility Normalized",
            "type": "string"
          },
          "credibility_original": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credibility Original"
          },
          "credibility_scheme": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credibility Scheme"
          },
          "excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excerpt"
          },
          "handling_code": {
            "default": "open",
            "title": "Handling Code",
            "type": "string"
          },
          "jurisdiction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jurisdiction"
          },
          "location_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Text"
          },
          "object_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Object Id"
          },
          "object_value": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Object Value"
          },
          "predicate": {
            "title": "Predicate",
            "type": "string"
          },
          "record_id": {
            "title": "Record Id",
            "type": "string"
          },
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          },
          "valid_from": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          },
          "valid_to": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid To"
          },
          "verification_status": {
            "default": "unverified",
            "title": "Verification Status",
            "type": "string"
          }
        },
        "required": [
          "subject_id",
          "predicate",
          "record_id"
        ],
        "title": "ClaimIn",
        "type": "object"
      },
      "ClaimOut": {
        "properties": {
          "analytic_confidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analytic Confidence"
          },
          "assertion_type": {
            "title": "Assertion Type",
            "type": "string"
          },
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "claim_id": {
            "title": "Claim Id",
            "type": "string"
          },
          "collection_method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collection Method"
          },
          "credibility_normalized": {
            "title": "Credibility Normalized",
            "type": "string"
          },
          "credibility_original": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credibility Original"
          },
          "credibility_scheme": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credibility Scheme"
          },
          "excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excerpt"
          },
          "handling_code": {
            "title": "Handling Code",
            "type": "string"
          },
          "location_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Text"
          },
          "object_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Object Id"
          },
          "object_value": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Object Value"
          },
          "ontology_version": {
            "title": "Ontology Version",
            "type": "string"
          },
          "predicate": {
            "title": "Predicate",
            "type": "string"
          },
          "record_id": {
            "title": "Record Id",
            "type": "string"
          },
          "recorded_at": {
            "format": "date-time",
            "title": "Recorded At",
            "type": "string"
          },
          "retracted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retracted At"
          },
          "retraction_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retraction Reason"
          },
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          },
          "valid_from": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          },
          "valid_to": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid To"
          },
          "verification_status": {
            "title": "Verification Status",
            "type": "string"
          }
        },
        "required": [
          "claim_id",
          "subject_id",
          "predicate",
          "object_id",
          "object_value",
          "assertion_type",
          "record_id",
          "excerpt",
          "collection_method",
          "credibility_scheme",
          "credibility_original",
          "credibility_normalized",
          "verification_status",
          "analytic_confidence",
          "valid_from",
          "valid_to",
          "recorded_at",
          "retracted_at",
          "retraction_reason",
          "handling_code",
          "case_id",
          "location_text",
          "ontology_version"
        ],
        "title": "ClaimOut",
        "type": "object"
      },
...
      "EntityDetail": {
        "properties": {
          "claims_by_predicate": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/ClaimOut"
              },
              "type": "array"
            },
            "title": "Claims By Predicate",
            "type": "object"
          },
          "entity": {
            "$ref": "#/components/schemas/EntityOut"
          }
        },
        "required": [
          "entity",
          "claims_by_predicate"
        ],
        "title": "EntityDetail",
        "type": "object"
      },
...
      "EvidenceIn": {
        "properties": {
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "content_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Hash"
          },
          "description": {
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "handling_code": {
            "default": "restricted",
            "title": "Handling Code",
            "type": "string"
          },
          "legal_basis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legal Basis"
          },
          "record_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Record Id"
          },
          "storage_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storage Uri"
          }
        },
        "required": [
          "description"
        ],
        "title": "EvidenceIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RelationIn": {
        "properties": {
          "relation": {
            "title": "Relation",
            "type": "string"
          },
          "to_claim": {
            "title": "To Claim",
            "type": "string"
          }
        },
        "required": [
          "to_claim",
          "relation"
        ],
        "title": "RelationIn",
        "type": "object"
      },
      "RetractIn": {
        "properties": {
          "reason": {
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "RetractIn",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Governed claims-based intelligence platform (speckit Phase 2).",
    "title": "Aegis API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/claims": {
      "post": {
        "operationId": "createClaim",
        "parameters": [
          {
            "description": "Reason for access",
            "in": "query",
            "name": "purpose",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Reason for access",
              "title": "Purpose"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Claim",
        "tags": [
          "claims"
        ]
      }
    },
    "/v1/claims/{claim_id}": {
      "get": {
        "operationId": "getClaim",
        "parameters": [
          {
            "in": "path",
            "name": "claim_id",
            "required": true,
            "schema": {
              "title": "Claim Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "asOf",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asof"
            }
          },
          {
            "description": "Reason for access",
            "in": "query",
            "name": "purpose",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Reason for access",
              "title": "Purpose"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Claim",
        "tags": [
          "claims"
        ]
      }
    },
...
    "/v1/claims/{claim_id}/relations": {
      "post": {
        "operationId": "linkClaim",
        "parameters": [
          {
            "in": "path",
            "name": "claim_id",
            "required": true,
            "schema": {
              "title": "Claim Id",
              "type": "string"
            }
          },
          {
            "description": "Reason for access",
            "in": "query",
            "name": "purpose",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Reason for access",
              "title": "Purpose"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelationIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Linkclaim",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Link Claim",
        "tags": [
          "claims"
        ]
      }
    },
    "/v1/claims/{claim_id}/retract": {
      "post": {
        "operationId": "retractClaim",
        "parameters": [
          {
            "in": "path",
            "name": "claim_id",
            "required": true,
            "schema": {
              "title": "Claim Id",
              "type": "string"
            }
          },
          {
            "description": "Reason for access",
            "in": "query",
            "name": "purpose",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Reason for access",
              "title": "Purpose"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetractIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Retract Claim",
        "tags": [
          "claims"
        ]
      }
    },
    "/v1/entities/{entity_id}": {
      "get": {
        "operationId": "getEntity",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "asOf",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asof"
            }
          },
          {
            "description": "Reason for access",
            "in": "query",
            "name": "purpose",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Reason for access",
              "title": "Purpose"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Entity",
        "tags": [
          "entities"
        ]
      }
    },
...
    "/v1/evidence": {
      "post": {
        "operationId": "registerEvidence",
        "parameters": [
          {
            "description": "Reason for access",
            "in": "query",
            "name": "purpose",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Reason for access",
              "title": "Purpose"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvidenceIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Register Evidence",
        "tags": [
          "evidence"
        ]
      }
    },
    "/v1/evidence/{evidence_id}": {
      "get": {
        "operationId": "getEvidence",
        "parameters": [
          {
            "in": "path",
            "name": "evidence_id",
            "required": true,
            "schema": {
              "title": "Evidence Id",
              "type": "string"
            }
          },
          {
            "description": "Reason for access",
            "in": "query",
            "name": "purpose",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Reason for access",
              "title": "Purpose"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Getevidence",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Evidence",
        "tags": [
          "evidence"
        ]
      }
    },
    "/v1/evidence/{evidence_id}/custody-events": {
      "post": {
        "operationId": "addCustodyEvent",
        "parameters": [
          {
            "in": "path",
            "name": "evidence_id",
            "required": true,
            "schema": {
              "title": "Evidence Id",
              "type": "string"
            }
          },
          {
            "description": "Reason for access",
            "in": "query",
            "name": "purpose",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Reason for access",
              "title": "Purpose"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustodyEventIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Addcustodyevent",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Custody Event",
        "tags": [
          "evidence"
        ]
      }
    },
...
  }
}